fix(scan): address review — share UpsertCache, lazy probe, symlink sidecars
- Move `UpsertCache` (artist/genre memo) out of the Tauri command into
  `waveflow_core::scanner` returning `CoreResult`, so the command only
  orchestrates and the reusable lookup logic lives next to
  `upsert_artist` / `upsert_genre` / `split_artist_name`. `scan.rs`
  imports the shared type.
- Build the insert-vs-update probe map only when `to_extract` is
  non-empty: a stat-skip rescan no longer pays a library-wide row load
  for nothing.
- Preserve symlinked sidecar images in `read_dir_artist_images`:
  `DirEntry::file_type()` doesn't follow links, so fall back to the
  link-following `Path::is_file` for symlink entries (regular files
  still skip the stat).
- Include the final `tx.commit()` in `db_us` so `db_ms_total` no longer
  omits the last batch's commit cost.
