+++ src-tauri/crates/core/src/scanner/mod.rs
@@ -33,7 +33,7 @@ pub use upserts::{
     link_local_artist_image, link_va_artist_image, maybe_link_artist_images,
     merge_implicit_compilations, now_millis, resolve_album_artist, split_artist_name, upsert_album,
     upsert_artist, upsert_artist_list, upsert_artwork, upsert_genre, ArtistImageScanCache,
-    VARIOUS_ARTISTS_LABEL,
+    UpsertCache, VARIOUS_ARTISTS_LABEL,
 };
 
 /// Helper used inside the audio file extractors. Pulled out into the
